requirements.txt
asn1crypto==0.22.0
beautifulsoup4==4.6.0
click==6.7
Django==1.11.2
django-bootstrap-form==3.3
future==0.16.0
Jinja2==2.9.6
line-bot-sdk==1.5.0
lxml==3.8.0
matplotlib==2.0.2
numpy==1.13.1
pandas==0.20.3
pip-tools==1.9.0
pipenv==7.2.6
prompt-toolkit==1.0.14
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.1
scikit-learn==0.19.0
six==1.10.0
gunicorn==19.7.1
dj-database-url==0.4.2
whitenoise==3.3.0
